Adds default background check
Checks to see if there are any files currently in the background folder, if there are not (i.e. the user deleted everything or first-time setup) then it will copy the defaults from the default_background folder. This will allow the user to replace the default backgrounds without having them added back to the folder with subsequent app updates.